+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/ir/TracingTest.kt
@@ -4,7 +4,6 @@ package dev.zacsweers.metro.compiler.ir
 
 import com.google.common.truth.Truth.assertThat
 import dev.zacsweers.metro.compiler.MetroCompilerTest
-import kotlin.io.path.listDirectoryEntries
 import kotlin.io.path.readText
 import org.junit.Test
 
@@ -37,7 +36,7 @@ class TracingTest : MetroCompilerTest() {
       ),
       options = metroOptions.copy(reportsDestination = reportsDir),
     ) {
-      val timings = reportsDir.listDirectoryEntries().single().resolve("timings.csv").readText()
+      val timings = reportsDir.resolve("timings.csv").readText()
       val withoutTime = timings.lines().drop(1).joinToString("\n") { it.substringBeforeLast(",") }
       assertThat(withoutTime)
         .isEqualTo(
@@ -72,7 +71,7 @@ class TracingTest : MetroCompilerTest() {
             .trimIndent()
         )
 
-      val traceLog = reportsDir.listDirectoryEntries().single().resolve("traceLog.txt").readText()
+      val traceLog = reportsDir.resolve("traceLog.txt").readText()
       val cleanedLog = traceLog.replace("\\((\\d+) ms\\)".toRegex(), "(xx ms)")
       assertThat(cleanedLog.trim())
         .isEqualTo(
@@ -169,7 +168,7 @@ class TracingTest : MetroCompilerTest() {
       ),
       options = metroOptions.copy(reportsDestination = reportsDir),
     ) {
-      val timings = reportsDir.listDirectoryEntries().single().resolve("timings.csv").readText()
+      val timings = reportsDir.resolve("timings.csv").readText()
       val withoutTime = timings.lines().drop(1).joinToString("\n") { it.substringBeforeLast(",") }
       assertThat(withoutTime)
         .isEqualTo(
@@ -227,7 +226,7 @@ class TracingTest : MetroCompilerTest() {
             .trimIndent()
         )
 
-      val traceLog = reportsDir.listDirectoryEntries().single().resolve("traceLog.txt").readText()
+      val traceLog = reportsDir.resolve("traceLog.txt").readText()
       val cleanedLog = traceLog.replace("\\((\\d+) ms\\)".toRegex(), "(xx ms)")
       assertThat(cleanedLog.trim())
         .isEqualTo(
@@ -370,7 +369,7 @@ class TracingTest : MetroCompilerTest() {
       ),
       options = metroOptions.copy(reportsDestination = reportsDir),
     ) {
-      val timings = reportsDir.listDirectoryEntries().single().resolve("timings.csv").readText()
+      val timings = reportsDir.resolve("timings.csv").readText()
       val withoutTime = timings.lines().drop(1).joinToString("\n") { it.substringBeforeLast(",") }
       assertThat(withoutTime)
         .isEqualTo(
@@ -428,7 +427,7 @@ class TracingTest : MetroCompilerTest() {
             .trimIndent()
         )
 
-      val traceLog = reportsDir.listDirectoryEntries().single().resolve("traceLog.txt").readText()
+      val traceLog = reportsDir.resolve("traceLog.txt").readText()
       val cleanedLog = traceLog.replace("\\((\\d+) ms\\)".toRegex(), "(xx ms)")
       assertThat(cleanedLog.trim())
         .isEqualTo(