+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/metroGradleUtil.kt
@@ -1,0 +1,9 @@
+// Copyright (C) 2025 Zac Sweers
+// SPDX-License-Identifier: Apache-2.0
+package dev.zacsweers.metro.gradle
+
+import java.util.Locale
+
+internal fun String.capitalizeUS() = replaceFirstChar {
+  if (it.isLowerCase()) it.titlecase(Locale.US) else it.toString()
+}